requirements.txt
pandas
tqdm
torchvision

scikit-learn
matplotlib
seaborn
